fixed bugs and removed redundant files
